requirements.txt
openai
os
json
python--dotenv
gradio
